requirements.txt
duckdb>=0.10.0
pandas>=2.0.0
requests>=2.31.0
groq>=0.5.0
google-generativeai>=0.5.0
